final refactor and add readme and step by step
